+++ internal/cmd/company/iam/list_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestList(t *testing.T) {
+func TestListAllIAMIdentities(t *testing.T) {
 	testCases := map[string]struct {
 		server       *httptest.Server
 		clientConfig *client.Config
@@ -81,9 +81,87 @@ func TestList(t *testing.T) {
 	}
 }
 
+func TestListUsersIdentities(t *testing.T) {
+	testCases := map[string]struct {
+		server       *httptest.Server
+		clientConfig *client.Config
+		companyID    string
+		searchParams map[string]bool
+		err          bool
+	}{
+		"valid get response": {
+			server:    mockListServer(t),
+			companyID: "success",
+			clientConfig: &client.Config{
+				Transport: http.DefaultTransport,
+			},
+			searchParams: map[string]bool{},
+		},
+		"invalid body response": {
+			server:    mockListServer(t),
+			companyID: "fail",
+			clientConfig: &client.Config{
+				Transport: http.DefaultTransport,
+			},
+			err: true,
+		},
+	}
+
+	for testName, testCase := range testCases {
+		t.Run(testName, func(t *testing.T) {
+			defer testCase.server.Close()
+			testCase.clientConfig.Host = testCase.server.URL
+			client, err := client.APIClientForConfig(testCase.clientConfig)
+			require.NoError(t, err)
+
+			err = listSpecificEntities(context.TODO(), client, testCase.companyID, UsersEntityName)
+			if testCase.err {
+				assert.Error(t, err)
+			} else {
+				assert.NoError(t, err)
+			}
+		})
+	}
+}
+
 func mockListServer(t *testing.T) *httptest.Server {
 	t.Helper()
-	validBodyString := `[
+
+	return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		params, err := url.ParseQuery(r.URL.RawQuery)
+		require.NoError(t, err)
+		switch {
+		case r.Method == http.MethodGet && r.URL.Path == fmt.Sprintf(listAllIAMEntitiesTemplate, "success"):
+			assert.Equal(t, 0, len(params))
+			w.WriteHeader(http.StatusOK)
+			w.Write([]byte(validListIAMIdentitiesBodyString))
+		case r.Method == http.MethodGet && r.URL.Path == fmt.Sprintf(listAllIAMEntitiesTemplate, "search"):
+			searchTerms, ok := params["identityType"]
+			assert.True(t, ok)
+			assert.ElementsMatch(t, []string{"group", "serviceAccount"}, searchTerms)
+			w.WriteHeader(http.StatusOK)
+			w.Write([]byte(filteredListIAMIdentitiesString))
+		case r.Method == http.MethodGet && r.URL.Path == fmt.Sprintf(listUserEntityTemplate, "success"):
+			assert.Equal(t, 0, len(params))
+			w.WriteHeader(http.StatusOK)
+			w.Write([]byte(validListUserIdentitiesBodyString))
+		case r.Method == http.MethodGet && r.URL.Path == fmt.Sprintf(listAllIAMEntitiesTemplate, "fail"):
+			assert.Equal(t, 0, len(params))
+			w.WriteHeader(http.StatusOK)
+			w.Write([]byte("invalid json"))
+		case r.Method == http.MethodGet && r.URL.Path == fmt.Sprintf(listUserEntityTemplate, "fail"):
+			assert.Equal(t, 0, len(params))
+			w.WriteHeader(http.StatusOK)
+			w.Write([]byte("invalid json"))
+		default:
+			w.WriteHeader(http.StatusNotFound)
+			require.Fail(t, "unsupported call")
+		}
+	}))
+}
+
+const (
+	validListIAMIdentitiesBodyString = `[
   {
     "identityId": "000000000000000000000000",
     "email": "user.email@example.com",
@@ -101,7 +179,6 @@ func mockListServer(t *testing.T) *httptest.Server {
     "membersCount": 1,
     "members": [
       {
-        "_id": "000000000000000000000000",
         "name": "User Complete Name",
         "email": "user.email@example.com"
       }
@@ -116,7 +193,7 @@ func mockListServer(t *testing.T) *httptest.Server {
     "authMethod": "client_secret_basic"
   }
 ]`
-	filteredString := `[
+	filteredListIAMIdentitiesString = `[
     {
       "identityId": "000000000000000000000001",
       "name": "Group Name",
@@ -126,7 +203,6 @@ func mockListServer(t *testing.T) *httptest.Server {
       "membersCount": 1,
       "members": [
         {
-          "_id": "000000000000000000000000",
           "name": "User Complete Name",
           "email": "user.email@example.com"
         }
@@ -141,26 +217,35 @@ func mockListServer(t *testing.T) *httptest.Server {
       "authMethod": "client_secret_basic"
     }
 ]`
-	return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		params, err := url.ParseQuery(r.URL.RawQuery)
-		require.NoError(t, err)
-		switch {
-		case r.Method == http.MethodGet && r.URL.Path == fmt.Sprintf(listAllIAMEntitiesTemplate, "success"):
-			assert.Equal(t, 0, len(params))
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(validBodyString))
-		case r.Method == http.MethodGet && r.URL.Path == fmt.Sprintf(listAllIAMEntitiesTemplate, "search"):
-			searchTerms, ok := params["identityType"]
-			assert.True(t, ok)
-			assert.ElementsMatch(t, []string{"group", "serviceAccount"}, searchTerms)
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte(filteredString))
-		case r.Method == http.MethodGet && r.URL.Path == fmt.Sprintf(listAllIAMEntitiesTemplate, "fail"):
-			w.WriteHeader(http.StatusOK)
-			w.Write([]byte("invalid json"))
-		default:
-			w.WriteHeader(http.StatusNotFound)
-			require.Fail(t, "unsupported call")
-		}
-	}))
-}
+	validListUserIdentitiesBodyString = `[
+  {
+    "userId": "000000000000000000000001",
+    "email": "user.email@example.com",
+    "fullName": "User Full Name",
+    "companyRoles": [],
+    "lastLogin": "2010-01-01T00:00:00.000Z",
+    "groups": [{
+      "name": "Role Name",
+      "roleId": "role-id"
+    }]
+  },
+  {
+    "userId": "000000000000000000000002",
+    "email": "user.email@example.com",
+    "fullName": "User Full Name",
+    "companyRoles": ["role-id"],
+    "lastLogin": "2010-01-01T00:00:00.000Z"
+  },
+  {
+    "userId": "000000000000000000000003",
+    "email": "user.email@example.com",
+    "fullName": "User Full Name",
+    "companyRoles": ["role-id"],
+    "lastLogin": "2010-01-01T00:00:00.000Z",
+    "groups": [{
+      "name": "Role Name",
+      "roleId": "role-id"
+    }]
+  }
+]`
+)